Move ssh code to a client
